Implement InitGenesis for x/stake (closes #737)
