+++ x/stake/test_common.go
@@ -2,6 +2,7 @@ package stake
 
 import (
 	"encoding/hex"
+	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -51,6 +52,31 @@ var (
 	emptyAddr   sdk.Address
 	emptyPubkey crypto.PubKey
 )
+
+// default params for testing
+func defaultParams() Params {
+	return Params{
+		InflationRateChange: sdk.NewRat(13, 100),
+		InflationMax:        sdk.NewRat(20, 100),
+		InflationMin:        sdk.NewRat(7, 100),
+		GoalBonded:          sdk.NewRat(67, 100),
+		MaxValidators:       100,
+		BondDenom:           "fermion",
+	}
+}
+
+// initial pool for testing
+func initialPool() Pool {
+	return Pool{
+		TotalSupply:       0,
+		BondedShares:      sdk.ZeroRat,
+		UnbondedShares:    sdk.ZeroRat,
+		BondedPool:        0,
+		UnbondedPool:      0,
+		InflationLastTime: 0,
+		Inflation:         sdk.NewRat(7, 100),
+	}
+}
 
 // XXX reference the common declaration of this function
 func subspace(prefix []byte) (start, end []byte) {
@@ -123,6 +149,13 @@ func createTestInput(t *testing.T, sender sdk.Address, isCheckTx bool, initCoins
 	)
 	ck := bank.NewCoinKeeper(accountMapper)
 	keeper := NewKeeper(ctx, cdc, keyStake, ck)
+	encoded, err := json.Marshal(GenesisState{initialPool(), defaultParams()})
+	if err != nil {
+		panic(err)
+	}
+	if err = keeper.InitGenesis(ctx, encoded); err != nil {
+		panic(err)
+	}
 
 	// fill all the addresses with some coins
 	for _, addr := range addrs {
